LayoutTests/media/video-currentTime-set-expected.txt

Test that setting currentTime changes the time, and that 'ended' event is fired in a reasonable amount of time

EVENT(load)
EXPECTED (video.currentTime == '0') OK
RUN(video.currentTime = video.duration - 0.2)
EXPECTED (video.currentTime.toFixed(2) == '8.20') OK
RUN(video.play())

EVENT(ended)
END OF TEST
